feat: 拆分出handle-photos-data func
